\name{air}
\alias{air}
\title{AIR-Index}
\description{ The deseasonalized All Indian Rainfall index}
\usage{data(air)
}
\format{A matrix containing a time series of 1596 observations.
}
\source{
  Updated version of:
  D.A. Mooley and B. Parthasarathy, Fluctuations in all-India summer
  monsoon rainfall during 1871-1978, Clim. Change 6, 287-301 (1984)
}
\keyword{datasets}
