\name{nao}
\alias{nao}
\title{NAO-Index}
\description{ The North Atlantic Oscillation Index defined as the
  surface pressure difference between Iceland and Gibraltar
}
\usage{data(nao)
}
\format{A matrix containing a time series of 2140 observations.
}
\source{
  Climate Research Unit (CRU)
  

}
\keyword{datasets}
