\name{nino3}
\alias{nino3}
\title{NINO3-Index}
\description{ The NINO3-index defined as the mean monthly temperature
  anomaly in the area of the eastern tropical pacific: 5S-5N; 150W-90W.
}
\usage{data(nino3)
}
\format{A matrix containing a time series of 1788 observations.
}
\source{
  See: 
  A. Kaplan and M.A. Cane and Y. Kushnir and A.C. Clement and
  M.B. Blumenthal and B. Rajagopalan, Analyses of global sea surface
  temperature 1856-1991, J. Geophys. Res. 103, 18567-18589 (1998)

}
\keyword{datasets}
