\name{readts}
\alias{readts}
\title{Load time series from file}
\usage{
  readts(file)
}
\arguments{
  \item{file}{a character string giving the name of the
  file to load.}
}
\value{
  Returns time series object
}
\description{
  This function loads a tab separated time series with one
  time column and one value column in ASCII-format into a R
  time series object.
}
\examples{
##
}
\author{
  D. Maraun
}
\seealso{
  \code{\link[stats]{ts}}, \code{\link{readmatrix}}
}
\keyword{file}

