\name{writematrix}
\alias{writematrix}
\title{Writes matrix to file}
\usage{
  writematrix(file, data, header = "# R Matrix")
}
\arguments{
  \item{file}{a character string giving the name of the
  file.}

  \item{data}{matrix to be written.}
}
\value{
  No value returned
}
\description{
  This function writes a tab separated matrix with M
  columns into a file in ASCII-format.
}
\examples{
##
}
\author{
  D. Maraun
}
\seealso{
  \code{\link{readts}}, \code{\link{readmatrix}}
}
\keyword{file}

